+++ CHANGELOG.md
@@ -21,7 +21,10 @@ Adding a requirement of a major version of a dependency is breaking a contract.
 Dropping a requirement of a major version of a dependency is a new contract.
 
 ## [Unreleased]
-[Unreleased]: https://github.com/atlassian/virtual-users/compare/release-3.13.2...master
+[Unreleased]: https://github.com/atlassian/virtual-users/compare/release-3.14.0...master
+
+## [3.14.0] - 2023-05-29
+[3.14.0]: https://github.com/atlassian/virtual-users/compare/release-3.13.2...release-3.14.0
 
 ### Added
 - Add `VirtualUserNodeResult.countVusPerNode` to replace the `test-results/nodes.csv` file convention.